+++ jest.config.js
@@ -1,3 +1,2 @@
 module.exports = {
-	testMatch: ['<rootDir>/test/index.js']
-}
+};
